[P4.30] merge render_chart.py into renderers — each renderer self-contained
Deleted pipeline/render_chart.py (948 lines). Each renderer now owns
its builder + private helpers. Shared code in renderers/helpers.py.
CLI: python -m renderers.helpers --template gantt

Closes #78
